refactor(repl): all commands do the file path prepend check
